refactor user creation
