+++ api/index.ts
@@ -1,3 +1,4 @@
 export * from './upload/upload.controller';
 export * from './user/user.controller';
 export * from './user/user.service';
+export * from './models';
